+++ src/include/finder.hpp
@@ -31,9 +31,15 @@ class Finder {
       const Program &p, Sequence &norm_seq,
       const std::vector<OeisSequence> &sequences);
 
-  std::pair<std::string, Program> checkProgram(Program program,
-                                               Program existing, bool is_new,
-                                               const OeisSequence &seq);
+  std::pair<std::string, Program> checkProgramFull(Program program,
+                                                   Program existing,
+                                                   bool is_new,
+                                                   const OeisSequence &seq);
+
+  std::pair<std::string, Program> checkProgramBasic(Program program,
+                                                    Program existing,
+                                                    bool is_new,
+                                                    const OeisSequence &seq);
 
   std::vector<std::unique_ptr<Matcher>> &getMatchers() { return matchers; }